little classes refactoring
