+++ core/src/main/kotlin/ru/solodovnikov/rxlocationmanager/transformers.kt
@@ -6,8 +6,6 @@ interface RxLocationTransformer<RX> {
 
 interface BasePermissionTransformer<RX> : RxLocationTransformer<RX> {
     interface PermissionCallback {
-        fun shouldShowRequestPermissionRationale(permission: String): Boolean
-
         fun requestPermissions(permissions: Array<String>)
     }
 }
