+++ rxlocationmanager-rxjava2/src/main/kotlin/ru/solodovnikov/rx2locationmanager/manager.kt
@@ -1,18 +1,13 @@
 package ru.solodovnikov.rx2locationmanager
 
-import android.Manifest
 import android.content.Context
-import android.content.pm.PackageManager
 import android.location.Location
 import android.location.LocationListener
-import android.os.Build
 import android.os.Bundle
 import io.reactivex.*
-import io.reactivex.Observable
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.functions.Function
 import io.reactivex.subjects.PublishSubject
-import java.util.*
 import java.util.concurrent.TimeoutException
 
 /**
@@ -22,6 +17,7 @@ class RxLocationManager internal constructor(context: Context,
                                              private val scheduler: Scheduler) : BaseRxLocationManager<Single<Location>, Maybe<Location>>(context) {
     constructor(context: Context) : this(context, AndroidSchedulers.mainThread())
 
+    private val permissionResult by lazy { PublishSubject.create<Pair<Array<out String>, IntArray>>() }
 
     /**
      * @return Result [Maybe] will not emit any value if location is null.
@@ -82,75 +78,21 @@ class RxLocationManager internal constructor(context: Context,
                     }
 
     override fun onRequestPermissionsResult(permissions: Array<out String>, grantResults: IntArray) {
+
         permissionResult.onNext(Pair(permissions, grantResults))
     }
 
-    companion object {
-        private val permissionResult by lazy { PublishSubject.create<Pair<Array<out String>, IntArray>>() }
+    fun permissionsSingleTransformer(context: Context,
+                                     callback: BasePermissionTransformer.PermissionCallback): BasePermissionTransformer<Single<Location>>
+            = PermissionRxSingleTransformer(context, permissionResult, callback)
 
-        fun checkPermissionsSingle(context: Context,
-                                   callback: BasePermissionTransformer.PermissionCallback): BasePermissionTransformer<Single<Location>>
-                = PermissionRxSingleTransformer(context, permissionResult, callback)
-
-        fun checkPermissionsMaybe(context: Context,
-                                  callback: BasePermissionTransformer.PermissionCallback): BasePermissionTransformer<Maybe<Location>>
-                = PermissionRxMaybeTransformer(context, permissionResult, callback)
-    }
+    fun permissionsMaybeTransformer(context: Context,
+                                    callback: BasePermissionTransformer.PermissionCallback): BasePermissionTransformer<Maybe<Location>>
+            = PermissionRxMaybeTransformer(context, permissionResult, callback)
 
     private fun applySchedulers(s: Single<Location>) = s.subscribeOn(scheduler)
 
     private fun applySchedulers(m: Maybe<Location>) = m.subscribeOn(scheduler)
-
-    abstract class BasePermissionTransformerImpl<RX>(context: Context,
-                                                     private val permissionResult: PublishSubject<Pair<Array<out String>, IntArray>>,
-                                                     private val callback: BasePermissionTransformer.PermissionCallback
-    ) : BasePermissionTransformer<RX> {
-        private val context = context.applicationContext
-
-        private val permissions =
-                arrayOf(Manifest.permission.ACCESS_FINE_LOCATION,
-                        Manifest.permission.ACCESS_COARSE_LOCATION)
-
-        protected fun checkPermissions(): Completable =
-                Completable.create { emitter ->
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                        val deniedP = permissions.filter {
-                            context.checkSelfPermission(it) == PackageManager.PERMISSION_DENIED
-                        }.toTypedArray()
-
-                        if (deniedP.isNotEmpty()) {
-                            callback.requestPermissions(deniedP)
-                            permissionResult.subscribe {
-                                val resultPermissions = it.first
-                                val resultPermissionsResults = it.second
-                                if (!Arrays.equals(resultPermissions, deniedP) || resultPermissionsResults.find { it == PackageManager.PERMISSION_DENIED } != null) {
-                                    emitter.onError(SecurityException("User denied permissions: ${deniedP.asList()}"))
-                                } else {
-                                    emitter.onComplete()
-                                }
-                            }.apply { emitter.setCancellable { dispose() } }
-                        } else {
-                            emitter.onComplete()
-                        }
-                    } else {
-                        emitter.onComplete()
-                    }
-                }
-    }
-
-    private class PermissionRxSingleTransformer(context: Context,
-                                                permissionResult: PublishSubject<Pair<Array<out String>, IntArray>>,
-                                                callback: BasePermissionTransformer.PermissionCallback
-    ) : BasePermissionTransformerImpl<Single<Location>>(context, permissionResult, callback) {
-        override fun transform(rx: Single<Location>): Single<Location> = checkPermissions().andThen(rx)
-    }
-
-    private class PermissionRxMaybeTransformer(context: Context,
-                                               permissionResult: PublishSubject<Pair<Array<out String>, IntArray>>,
-                                               callback: BasePermissionTransformer.PermissionCallback
-    ) : BasePermissionTransformerImpl<Maybe<Location>>(context, permissionResult, callback) {
-        override fun transform(rx: Maybe<Location>): Maybe<Location> = checkPermissions().andThen(rx)
-    }
 }
 
 /**
