+++ sample-rxjava2-kotlin/src/main/kotlin/ru/solodovnikov/rxlocationmanager/sample/MainActivity.kt
@@ -6,15 +6,12 @@ import android.os.Bundle
 import android.support.design.widget.CoordinatorLayout
 import android.support.design.widget.Snackbar
 import android.support.v4.app.ActivityCompat
-import android.support.v4.app.Fragment
-import android.support.v4.content.ContextCompat
 import android.support.v7.app.AppCompatActivity
 import android.support.v7.widget.Toolbar
 import android.view.Menu
 import android.view.MenuItem
 import io.reactivex.Maybe
 import io.reactivex.Single
-import io.reactivex.subjects.PublishSubject
 import ru.solodovnikov.rx2locationmanager.*
 import java.util.concurrent.TimeUnit
 
@@ -76,7 +73,7 @@ class MainActivity : AppCompatActivity(), BasePermissionTransformer.PermissionCa
     }
 
     private fun requestLastNetworkLocation() {
-        rxLocationManager.getLastLocation(LocationManager.NETWORK_PROVIDER)
+        rxLocationManager.getLastLocation(LocationManager.NETWORK_PROVIDER, transformers = rxLocationManager.permissionsMaybeTransformer(this, this))
                 .testSubscribe("requestLastNetworkLocation")
     }
